+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+  - Every record managed by External DNS is now mapped to a kubernetes resource (service/ingress) @ideahitme
+    - New field is stored in TXT DNS record which reflects which kubernetes resource has acquired the DNS name
+    - Target of DNS record is changed only if corresponding kubernetes resource target changes
+    - If kubernetes resource is deleted, then another resource may acquire DNS name
+    - "Flapping" target issue is resolved by providing a consistent and defined mechanism for choosing a target
+      
+## v0.4.8 - 2017-11-22
+
+  - Allow filtering by source annotation via `--annotation-filter` (#354) @khrisrichardson
+  - Add support for Headless hostPort services (#324)
+  - AWS: Added change batch limiting to a maximum of 4000 Route53 updates in one API call.  Changes exceeding the limit will be dropped but all related changes by hostname are preserved within the limit. (#368) @bitvector2
+  - Google: Support configuring TTL by annotation: `external-dns.alpha.kubernetes.io/ttl`. (#389) @stealthybox
+  - Infoblox: add option `--no-infoblox-ssl-verify` (#378)
+  - Inmemory: add support to specify zones for inmemory provider via command line (#366)
+
 ## v0.4.7 - 2017-10-18
 
   - CloudFlare: Disable proxy mode for TXT and others (#361) @dunglas